# Improves uuid writer perf by making use of Utf8Formatter