This PR improves Uuid writer by making use of [Utf8Formatter.TryFormat](https://learn.microsoft.com/en-us/dotnet/api/system.buffers.text.utf8formatter.tryformat?view=net-6.0#system-buffers-text-utf8formatter-tryformat) which Formats a `Guid` as a UTF8 string and allows to write bytes directly to a destination `Span`.

| Method    | Mean     | Error    | StdDev   | Ratio | Gen0   | Allocated | Alloc Ratio |
|---------- |---------:|---------:|---------:|------:|-------:|----------:|------------:|
| Current   | 47.04 ns | 0.309 ns | 0.289 ns |  1.00 | 0.0306 |     128 B |        1.00 |
| Optimized | 30.12 ns | 0.102 ns | 0.095 ns |  0.64 | 0.0076 |      32 B |        0.25 |